+++ tests/test_prg_engine_runtime_surface_functions_tests.h
@@ -603,6 +603,7 @@ namespace copperfin::runtime_surface_tests
     void test_same_prg_native_bindevent_property_access_and_assign_dispatch_preserve_current_event_metadata();
     void test_same_prg_native_bindevent_property_handler_fault_does_not_disable_future_dispatch();
     void test_property_read_handler_releasing_source_fails_catchably();
+    void test_method_before_handler_releasing_source_fails_catchably();
     void test_same_prg_native_access_assign_methods_virtualize_ordinary_property_reads_and_writes();
     void test_native_assigner_same_property_write_uses_raw_storage();
     void test_native_accessor_backed_properties_reflect_through_getpem_pemstatus_and_amembers();